+++ src/api/endpoints/locations.api.ts
@@ -1,5 +1,5 @@
 import { API, apiRoutes } from '..';
-import { ILocation, ILocationResponse } from '../types/locations';
+import { ILocation, ILocationResponse } from '../types/location';
 
 export const getLocations = async (): Promise<ILocation[]> => {
     const response: ILocationResponse = await API.get(apiRoutes.location);
