--- src/api/types/locations.ts
@@ -1,8 +1,0 @@
-export interface ILocationResponse {
-    locations: ILocation[];
-}
-
-export interface ILocation {
-    id: string;
-    name: string;
-}
