+++ src/types/index.ts
@@ -1,3 +1,3 @@
-export * from "./trip";
+export * from './trip';
 export * from './language';
- 
+export * from './user';
