+++ src/types/location.d.ts
@@ -1,0 +1,4 @@
+export interface ILocation {
+    id: string;
+    name: string;
+}
