--- src/types/trip.ts
@@ -1,3 +1,0 @@
-export interface Trip {
-  id: string;
-}
